mainline ethernet
